[tool] upgrade convert_hf_to_torch_dist to support
