--- slime/backends/megatron_utils/models/__init__.py
@@ -1,108 +1,0 @@
-# Adapt from https://github.com/NVIDIA/Megatron-LM/blob/b1efb3c7126ef7615e8c333432d76e08038e17ff/pretrain_gpt.py
-import inspect
-from contextlib import nullcontext
-
-from megatron.core.enums import ModelType
-from megatron.core.models.gpt import GPTModel
-from megatron.core.models.gpt.gpt_layer_specs import (
-    get_gpt_decoder_block_spec,
-    get_gpt_layer_local_spec,
-    get_gpt_layer_with_transformer_engine_spec,
-)
-from megatron.core.transformer.spec_utils import import_module
-from megatron.training import get_args
-from megatron.training.arguments import core_transformer_config_from_args
-
-
-def model_provider(pre_process=True, post_process=True) -> GPTModel:
-    """Builds the model.
-
-    If you set the use_legacy_models to True, it will return the legacy GPT model and if not the mcore GPT model.
-
-    Args:
-        pre_process (bool, optional): Set to true if you need to compute embedings. Defaults to True.
-        post_process (bool, optional): Set to true if you need to want to compute output logits/loss. Defaults to True.
-
-
-    Returns:
-        Union[GPTModel, megatron.legacy.model.GPTModel]: The returned model
-    """
-    args = get_args()
-    use_te = args.transformer_impl == "transformer_engine"
-
-    # Experimental loading arguments from yaml
-    config = core_transformer_config_from_args(args)
-
-    if args.spec is not None:
-        transformer_layer_spec = import_module(args.spec)
-        # Allow the spec to be a function so that user can use customized Megatron easier.
-        if callable(transformer_layer_spec):
-            transformer_layer_spec = transformer_layer_spec(args)
-    else:
-        if args.num_experts:
-            # Define the decoder block spec
-            transformer_layer_spec = get_gpt_decoder_block_spec(config, use_transformer_engine=use_te)
-        else:
-            # Define the decoder layer spec
-            if use_te:
-                transformer_layer_spec = get_gpt_layer_with_transformer_engine_spec(
-                    args.num_experts,
-                    args.moe_grouped_gemm,
-                    args.qk_layernorm,
-                    args.multi_latent_attention,
-                    args.moe_use_legacy_grouped_gemm,
-                )
-            else:
-                transformer_layer_spec = get_gpt_layer_local_spec(
-                    args.num_experts,
-                    args.moe_grouped_gemm,
-                    args.qk_layernorm,
-                    args.multi_latent_attention,
-                    args.moe_use_legacy_grouped_gemm,
-                )
-
-    build_model_context = nullcontext
-    build_model_context_args = {}
-    if args.fp8_param_gather:
-        try:
-            from transformer_engine.pytorch import fp8_model_init
-
-            build_model_context = fp8_model_init
-            build_model_context_args["enabled"] = True
-
-            # Check if fp8_model_init supports preserve_high_precision_init_val
-            if "preserve_high_precision_init_val" in inspect.signature(fp8_model_init).parameters:
-                build_model_context_args["preserve_high_precision_init_val"] = True
-        except:
-            raise RuntimeError("--fp8-param-gather requires `fp8_model_init` from TransformerEngine, but not found.")
-
-    kwargs = {
-        "config": config,
-        "transformer_layer_spec": transformer_layer_spec,
-        "vocab_size": args.padded_vocab_size,
-        "max_sequence_length": args.max_position_embeddings,
-        "pre_process": pre_process,
-        "post_process": post_process,
-        "fp16_lm_cross_entropy": args.fp16_lm_cross_entropy,
-        "parallel_output": True,
-        "share_embeddings_and_output_weights": not args.untie_embeddings_and_output_weights,
-        "position_embedding_type": args.position_embedding_type,
-        "rotary_percent": args.rotary_percent,
-        "rotary_base": args.rotary_base,
-        "rope_scaling": args.use_rope_scaling,
-    }
-
-    if getattr(args, "mtp_num_layers", None):
-        from megatron.core.models.gpt.gpt_layer_specs import get_gpt_mtp_block_spec
-
-        mtp_block_spec = get_gpt_mtp_block_spec(config, transformer_layer_spec, use_transformer_engine=use_te)
-        kwargs["mtp_block_spec"] = mtp_block_spec
-
-    with build_model_context(**build_model_context_args):
-        model = GPTModel(**kwargs)
-
-    return model
-
-
-def get_model_provider_and_type():
-    return model_provider, ModelType.encoder_or_decoder
